pyproject.toml
[tool.poetry]
name = "llm-mastodon-summarizer"
version = "0.1.0"
description = ""
authors = ["Lawhacknifemi <lawalsheriffnifemi@gmail.com>"]
readme = "README.md"
#packages = [{include = "llm_mastodon_summarizer"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
openai = "^0.27.2"
python-dotenv = "^0.21.1"
tiktoken = "^0.2.0"
streamlit = "^1.18.1"
pyrate-limiter = "^2.10.0"
praw = "^7.7.0"
colorlog = "^6.7.0"
anthropic = "^0.3.2"
flask = "^2.3.2"
mastodon-py = "^1.8.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
mypy = "^0.991"
types-requests = "^2.28.11.8"
black = "^23.1.0"
debugpy = "^1.6.6"
pylint = "^2.16.2"

[tool.pylint]
init-hook='import sys; sys.path.append("./app")'
disable = [
    "E0402"
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
